swupdate.bbclass: fix padding to prevent SIGPIPE
cpio seems to add zero padding to the swu file up to the block size. swupdate
expects it to padding till it is 4 bytes aligned again, see [1]. These trailing
zeros can crash swupdate by causing a SIGPIPE. So make sure only the trailing
zeros as expected by swupdate are in the swu file.

[1] https://github.com/sbabic/swupdate/blob/2016.10/core/cpio_utils.c#L322